==> src/Competencia2/UltimoIntento/2_FE_inicial.ipynb <==
#limpio la memoria
rm( list=ls() )  #remove all objects
gc()             #garbage collection

require("data.table")

#Parametros del script
#setwd("C:/_MCD/Labo1/code")  #Establezco el Working Directory
setwd("~/buckets/b1/")

#cargo el dataset
dataset  <- fread("./datasets/alt_c2_reparacNA.csv.gz")   #donde entreno# FIN Parametros del script


AgregarVariables  <- function( dataset )
{
  gc()
  #INICIO de la seccion donde se deben hacer cambios con variables nuevas

  #creo un ctr_quarter que tenga en cuenta cuando los clientes hace 3 menos meses que estan
  dataset[  , ctrx_quarter_normalizado := ctrx_quarter ]
  dataset[ cliente_antiguedad==1 , ctrx_quarter_normalizado := ctrx_quarter * 5 ]
  dataset[ cliente_antiguedad==2 , ctrx_quarter_normalizado := ctrx_quarter * 2 ]
  dataset[ cliente_antiguedad==3 , ctrx_quarter_normalizado := ctrx_quarter * 1.2 ]
  
  #variable extraida de una tesis de maestria de Irlanda
  dataset[  , mpayroll_sobre_edad  := mpayroll / cliente_edad ]

  #se crean los nuevos campos para MasterCard  y Visa, teniendo en cuenta los NA's
  #varias formas de combinar Visa_status y Master_status
  dataset[ , vm_status01       := pmax( Master_status,  Visa_status, na.rm = TRUE) ]
  dataset[ , vm_status02       := Master_status +  Visa_status ]
  dataset[ , vm_status03       := pmax( ifelse( is.na(Master_status), 10, Master_status) , ifelse( is.na(Visa_status), 10, Visa_status) ) ]
  dataset[ , vm_status04       := ifelse( is.na(Master_status), 10, Master_status)  +  ifelse( is.na(Visa_status), 10, Visa_status)  ]
  dataset[ , vm_status05       := ifelse( is.na(Master_status), 10, Master_status)  +  100*ifelse( is.na(Visa_status), 10, Visa_status)  ]

  dataset[ , vm_status06       := ifelse( is.na(Visa_status), 
                                          ifelse( is.na(Master_status), 10, Master_status), 
                                          Visa_status)  ]

  dataset[ , mv_status07       := ifelse( is.na(Master_status), 
                                          ifelse( is.na(Visa_status), 10, Visa_status), 
                                          Master_status)  ]


  #combino MasterCard y Visa
  dataset[ , vm_mfinanciacion_limite := rowSums( cbind( Master_mfinanciacion_limite,  Visa_mfinanciacion_limite) , na.rm=TRUE ) ]

  dataset[ , vm_Fvencimiento         := pmin( Master_Fvencimiento, Visa_Fvencimiento, na.rm = TRUE) ]
  dataset[ , vm_Finiciomora          := pmin( Master_Finiciomora, Visa_Finiciomora, na.rm = TRUE) ]
  dataset[ , vm_msaldototal          := rowSums( cbind( Master_msaldototal,  Visa_msaldototal) , na.rm=TRUE ) ]
  dataset[ , vm_msaldopesos          := rowSums( cbind( Master_msaldopesos,  Visa_msaldopesos) , na.rm=TRUE ) ]
  dataset[ , vm_msaldodolares        := rowSums( cbind( Master_msaldodolares,  Visa_msaldodolares) , na.rm=TRUE ) ]
  dataset[ , vm_mconsumospesos       := rowSums( cbind( Master_mconsumospesos,  Visa_mconsumospesos) , na.rm=TRUE ) ]
  dataset[ , vm_mconsumosdolares     := rowSums( cbind( Master_mconsumosdolares,  Visa_mconsumosdolares) , na.rm=TRUE ) ]
  dataset[ , vm_mlimitecompra        := rowSums( cbind( Master_mlimitecompra,  Visa_mlimitecompra) , na.rm=TRUE ) ]
  dataset[ , vm_madelantopesos       := rowSums( cbind( Master_madelantopesos,  Visa_madelantopesos) , na.rm=TRUE ) ]
  dataset[ , vm_madelantodolares     := rowSums( cbind( Master_madelantodolares,  Visa_madelantodolares) , na.rm=TRUE ) ]
  dataset[ , vm_fultimo_cierre       := pmax( Master_fultimo_cierre, Visa_fultimo_cierre, na.rm = TRUE) ]
  dataset[ , vm_mpagado              := rowSums( cbind( Master_mpagado,  Visa_mpagado) , na.rm=TRUE ) ]
  dataset[ , vm_mpagospesos          := rowSums( cbind( Master_mpagospesos,  Visa_mpagospesos) , na.rm=TRUE ) ]
  dataset[ , vm_mpagosdolares        := rowSums( cbind( Master_mpagosdolares,  Visa_mpagosdolares) , na.rm=TRUE ) ]
  dataset[ , vm_fechaalta            := pmax( Master_fechaalta, Visa_fechaalta, na.rm = TRUE) ]
  dataset[ , vm_mconsumototal        := rowSums( cbind( Master_mconsumototal,  Visa_mconsumototal) , na.rm=TRUE ) ]
  dataset[ , vm_cconsumos            := rowSums( cbind( Master_cconsumos,  Visa_cconsumos) , na.rm=TRUE ) ]
  dataset[ , vm_cadelantosefectivo   := rowSums( cbind( Master_cadelantosefectivo,  Visa_cadelantosefectivo) , na.rm=TRUE ) ]
  dataset[ , vm_mpagominimo          := rowSums( cbind( Master_mpagominimo,  Visa_mpagominimo) , na.rm=TRUE ) ]

  #a partir de aqui juego con la suma de Mastercard y Visa
  dataset[ , vmr_Master_mlimitecompra:= Master_mlimitecompra / vm_mlimitecompra ]
  dataset[ , vmr_Visa_mlimitecompra  := Visa_mlimitecompra / vm_mlimitecompra ]
  dataset[ , vmr_msaldototal         := vm_msaldototal / vm_mlimitecompra ]
  dataset[ , vmr_msaldopesos         := vm_msaldopesos / vm_mlimitecompra ]
  dataset[ , vmr_msaldopesos2        := vm_msaldopesos / vm_msaldototal ]
  dataset[ , vmr_msaldodolares       := vm_msaldodolares / vm_mlimitecompra ]
  dataset[ , vmr_msaldodolares2      := vm_msaldodolares / vm_msaldototal ]
  dataset[ , vmr_mconsumospesos      := vm_mconsumospesos / vm_mlimitecompra ]
  dataset[ , vmr_mconsumosdolares    := vm_mconsumosdolares / vm_mlimitecompra ]
  dataset[ , vmr_madelantopesos      := vm_madelantopesos / vm_mlimitecompra ]
  dataset[ , vmr_madelantodolares    := vm_madelantodolares / vm_mlimitecompra ]
  dataset[ , vmr_mpagado             := vm_mpagado / vm_mlimitecompra ]
  dataset[ , vmr_mpagospesos         := vm_mpagospesos / vm_mlimitecompra ]
  dataset[ , vmr_mpagosdolares       := vm_mpagosdolares / vm_mlimitecompra ]
  dataset[ , vmr_mconsumototal       := vm_mconsumototal  / vm_mlimitecompra ]
  dataset[ , vmr_mpagominimo         := vm_mpagominimo  / vm_mlimitecompra ]

  #Aqui debe usted agregar sus propias nuevas variables
     dataset[ , prop_vida_en_banco          := cliente_antiguedad/(12*cliente_edad)]
    
    #dataset[ , mas_una_caja_ahorro         := ifelse(ccaja_ahorro > 1,1,0)]
    
    #Prestamos
    dataset[ , cprestamos                 := rowSums( cbind( cprestamos_personales, cprestamos_prendarios, cprestamos_hipotecarios) , na.rm=TRUE ) ]
    dataset[ , mprestamos                 := rowSums( cbind( mprestamos_personales, mprestamos_prendarios, mprestamos_hipotecarios) , na.rm=TRUE ) ]
    dataset[ , tiene_prestamos             := ifelse(cprestamos > 0,1,0)]
    
    #Plazos Fijos
    dataset[ , mplazo_fijo                := rowSums( cbind( mplazo_fijo_pesos, mplazo_fijo_dolares) , na.rm=TRUE ) ]
    #dataset[ , tiene_plazos_fijos             := ifelse(cplazo_fijo > 0,1,0)]

    #Inversiones
    dataset[ , minversion1                := rowSums( cbind( minversion1_pesos, minversion1_dolares) , na.rm=TRUE ) ]
    dataset[ , cinversion12                := rowSums( cbind( cinversion1, cinversion2) , na.rm=TRUE ) ]
    dataset[ , minversion12                := rowSums( cbind( minversion1, minversion2) , na.rm=TRUE ) ]
    dataset[ , cinversion_total                := rowSums( cbind( cinversion12, cplazo_fijo) , na.rm=TRUE ) ]
    dataset[ , minversion_total                := rowSums( cbind( minversion12, mplazo_fijo) , na.rm=TRUE ) ]
    #dataset[ , tiene_inversiones_total             := ifelse(cinversion_total > 0,1,0)]
    dataset[ , tipo_inversion1_sobre_2             := ifelse(minversion2 == 0,0,minversion1/minversion2)]
    dataset[ , tipo_inversion_pf_sobre_total             := ifelse(minversion_total == 0,0,mplazo_fijo/minversion_total)]

    #Seguros
    dataset[ , cseguro                := rowSums( cbind( cseguro_vida, cseguro_auto,cseguro_vivienda,cseguro_accidentes_personales) , na.rm=TRUE ) ]
    #dataset[ , tiene_seguros             := ifelse(cseguro > 0,1,0)]

    #Debitos automaticos
    dataset[ , cdebitos_automaticos                := rowSums( cbind( ccuenta_debitos_automaticos,ctarjeta_visa_debitos_automaticos,ctarjeta_master_debitos_automaticos ) , na.rm=TRUE ) ]
    dataset[ , mdebitos_automaticos                := rowSums( cbind( mcuenta_debitos_automaticos,mttarjeta_visa_debitos_automaticos,mttarjeta_master_debitos_automaticos ) , na.rm=TRUE ) ]
    #dataset[ , tiene_debitos_automaticos              := ifelse(cdebitos_automaticos > 0,1,0)]

    #Servicios
    dataset[ , cpago_servicios_total                := rowSums( cbind( cpagodeservicios,cpagomiscuentas ) , na.rm=TRUE ) ]
    dataset[ , mpago_servicios_total                := rowSums( cbind( mpagodeservicios,mpagomiscuentas ) , na.rm=TRUE ) ]
    #dataset[ , paga_servicios             := ifelse(cpago_servicios_total > 0,1,0)]

    #Descuentos
    #dataset[ , cdescuentos                := rowSums( cbind( ccajeros_propios_descuentos,ctarjeta_visa_descuentos,ctarjeta_master_descuentos ) , na.rm=TRUE ) ]
    #dataset[ , mdescuentos                := rowSums( cbind( mcajeros_propios_descuentos,mtarjeta_visa_descuentos,mtarjeta_master_descuentos ) , na.rm=TRUE ) ]
    #dataset[ , tiene_descuentos             := ifelse(cdescuentos > 0,1,0)]

    #Comisiones
    dataset[ , ccomisiones                := rowSums( cbind( ccomisiones_mantenimiento,ccomisiones_otras ) , na.rm=TRUE ) ]
    dataset[ , mcomisiones                := rowSums( cbind( mcomisiones_mantenimiento,mcomisiones_otras ) , na.rm=TRUE ) ]
    #dataset[ , paga_comisiones             := ifelse(ccomisiones > 0,1,0)]

    #Moneda extranjera
    #dataset[ , opera_moneda_extranjera             := ifelse(cforex > 0,1,0)]

    #Receptor frecuente
    #dataset[ , receptor_transf_frecuente             := ifelse(mtransferencias_recibidas > 5,1,0)] ########## VERIFICAR PUNTO CORTE
    #dataset[ , receptor_cheque_frecuente             := ifelse(ccheques_depositados > 5,1,0)] ########## VERIFICAR PUNTO CORTE

    #Autoservicio
    #dataset[ , usa_autoservicio             := ifelse(cextraccion_autoservicio > 0,1,0)]

    #Cheques
    dataset[ , ccheques                := rowSums( cbind( ccheques_depositados,ccheques_emitidos ) , na.rm=TRUE ) ]
    dataset[ , mcheques                := rowSums( cbind( mcheques_depositados,mcheques_emitidos ) , na.rm=TRUE ) ]
    #dataset[ , usa_cheques             := ifelse(ccheques > 0,1,0)]
    #dataset[ , ccheques_rechazados                := rowSums( cbind(ccheques_depositados_rechazados,ccheques_emitidos_rechazados ) , na.rm=TRUE ) ]
    #dataset[ , con_cheques_rechazados             := ifelse(ccheques_rechazados > 0,1,0)]

    #Cajeros
    dataset[ , catm_total_trx                := rowSums( cbind(catm_trx,catm_trx_other ) , na.rm=TRUE ) ]
    dataset[ , matm_total_trx                := rowSums( cbind(matm,matm_other ) , na.rm=TRUE ) ]
    
    dataset[ , ccualquier_cajero_trx                := rowSums( cbind(catm_total_trx,cextraccion_autoservicio ) , na.rm=TRUE ) ]
    dataset[ , mcualquier_cajero_trx                := rowSums( cbind(matm_total_trx,mextraccion_autoservicio ) , na.rm=TRUE ) ]
    #dataset[ , usa_cualquier_cajero             := ifelse(ccualquier_cajero_trx > 0,1,0)]

    
    #Interacciones banco
     dataset[ , cinteracciones_banco_ult_mes    := rowSums( cbind(chomebanking_transacciones,ccajas_transacciones,ccallcenter_transacciones,ccualquier_cajero_trx ) , na.rm=TRUE ) ]
    
     dataset[ , prop_homeb_total                := ifelse(cinteracciones_banco_ult_mes==0,0,chomebanking_transacciones/cinteracciones_banco_ult_mes) ]
     dataset[ , prop_cajas_total                := ifelse(cinteracciones_banco_ult_mes==0,0,ccajas_transacciones/cinteracciones_banco_ult_mes) ]
     dataset[ , prop_callcenter_total           := ifelse(cinteracciones_banco_ult_mes==0,0,ccallcenter_transacciones/cinteracciones_banco_ult_mes) ]
     dataset[ , prop_cajero_total                  := ifelse(cinteracciones_banco_ult_mes==0,0,ccualquier_cajero_trx/cinteracciones_banco_ult_mes) ]

     dataset[ , prop_interac_mov                  := ifelse(ctrx_quarter_normalizado==0,0,cinteracciones_banco_ult_mes/ctrx_quarter_normalizado) ]
    #dataset[ , sin_interac_ult_mes             := ifelse(cinteracciones_banco_ult_mes > 0,1,0)]
    #dataset[ , sin_mov_ult_cuat             := ifelse(ctrx_quarter_normalizado > 0,1,0)]

    #Payroll
   dataset[ , cpayroll_total    := rowSums( cbind(cpayroll_trx,cpayroll2_trx ) , na.rm=TRUE ) ]
   dataset[ , mpayroll_total    := rowSums( cbind(mpayroll,mpayroll2 ) , na.rm=TRUE ) ]
    #dataset[ , acredita_sueldo  := ifelse(cpayroll_total > 0,1,0)]

    #Rentabilidad y productos
    dataset[ , mmargen    := rowSums( cbind(mactivos_margen,mpasivos_margen ) , na.rm=TRUE ) ]

    dataset[ , prop_rent_mes_anual                  := ifelse(mrentabilidad_annual==0,0,mrentabilidad/mrentabilidad_annual) ]
    dataset[ , prop_rent_mes_antiguedad                  := ifelse(cliente_antiguedad==0,0,mrentabilidad/cliente_antiguedad) ]
    dataset[ , prop_rent_mes_cprod                  := ifelse(cproductos==0,0,mrentabilidad/cproductos) ]
    dataset[ , prop_rent_anual_cprod                  := ifelse(cproductos==0,0,mrentabilidad_annual/cproductos) ]
    dataset[ , prop_comisiones_cprod                  := ifelse(cproductos==0,0,mcomisiones/cproductos) ]
    dataset[ , prop_comisiones_rent_mes                  := ifelse(mrentabilidad==0,0,mcomisiones/mrentabilidad) ]
    dataset[ , prop_margen_cprod                  := ifelse(cproductos==0,0,mmargen/cproductos) ]
    dataset[ , prop_margen_rent_mes                  := ifelse(mrentabilidad==0,0,mmargen/mrentabilidad) ]
    
    #dataset[ , paga_comisiones  := ifelse(mcomisiones > 0,1,0)]
    #dataset[ , da_margen  := ifelse(mmargen > 0,1,0)]
    #dataset[ , es_rentable  := ifelse(mrentabilidad > 0,1,0)]

    #Debito y credito
    dataset[ , ctarjeta_credito    := rowSums( cbind(ctarjeta_visa,ctarjeta_master ) , na.rm=TRUE ) ]
    #dataset[ , tiene_tarjeta_credito  := ifelse(ctarjeta_credito > 0,1,0)]
    dataset[ , ctarjeta_credito_transacciones    := rowSums( cbind(ctarjeta_visa_transacciones,ctarjeta_master_transacciones ) , na.rm=TRUE ) ]
    dataset[ , mtarjeta_credito_consumos    := rowSums( cbind(mtarjeta_visa_consumo,mtarjeta_master_consumo ) , na.rm=TRUE ) ]

    #dataset[ , usa_tarjeta_credito  := ifelse(ctarjeta_credito_transacciones > 0,1,0)]
    #dataset[ , usa_tarjeta_debito  := ifelse(ctarjeta_debito_transacciones > 0,1,0)]

    dataset[ , prop_tarj_debito_credito    := ifelse(ctarjeta_credito_transacciones==0,0,ctarjeta_debito_transacciones/ctarjeta_credito_transacciones) ]

    #Saldo
    dataset[ , prop_dolares_saldo   := ifelse(mcuentas_saldo==0,0,mcaja_ahorro_dolares/mcuentas_saldo) ]
    dataset[ , prop_plazos_fijos_saldo   := ifelse(mcuentas_saldo==0,0,mplazo_fijo/mcuentas_saldo) ]
    dataset[ , prop_inversion12_saldo   := ifelse(mcuentas_saldo==0,0,minversion12/mcuentas_saldo) ]
    dataset[ , prop_consumos_tarj_cred_saldo    := ifelse(mcuentas_saldo==0,0,mtarjeta_credito_consumos/mcuentas_saldo) ]
    dataset[ , prop_rent_mes_saldo   := ifelse(mcuentas_saldo==0,0,mrentabilidad/mcuentas_saldo) ]

    dataset[ , mcuentas_saldo_mas_invers    := rowSums( cbind(mcuentas_saldo,minversion_total ) , na.rm=TRUE ) ]
    dataset[ , mcuentas_saldo_mas_invers_mas_caja_seg    := ifelse(ccaja_seguridad==1,1000000,0) + mcuentas_saldo_mas_invers ]
    dataset[ , mgasto_mensual_estim    := rowSums( cbind(mdebitos_automaticos,mpago_servicios_total,mtransferencias_emitidas,mextraccion_autoservicio,mcheques_emitidos,ccajas_extracciones,mautoservicio,mtarjeta_credito_consumos ) , na.rm=TRUE ) ]
    dataset[ , mingreso_mensual_estim    := rowSums( cbind(mtransferencias_recibidas,mcheques_depositados,ccajas_depositos,mpayroll_total ) , na.rm=TRUE ) ]
    
    dataset[ , prop_gasto_saldo   := ifelse(mcuentas_saldo==0,0,mgasto_mensual_estim/mcuentas_saldo) ]
    dataset[ , prop_ingreso_saldo   := ifelse(mcuentas_saldo==0,0,mingreso_mensual_estim/mcuentas_saldo) ]

    #Visa y Master
    dataset[ , prop_limite_compra_saldo   := ifelse(mcuentas_saldo==0,0,mv_mlimitecompra/mcuentas_saldo) ]
    dataset[ , prop_limite_compra_gastos   := ifelse(mgasto_mensual_estim==0,0,mv_mlimitecompra/mgasto_mensual_estim) ]
    dataset[ , prop_limite_compra_ingresos   := ifelse(mingreso_mensual_estim==0,0,mv_mlimitecompra/mingreso_mensual_estim) ]
    dataset[ , prop_limite_compra_saldo_estim   := ifelse(mcuentas_saldo_mas_invers_mas_caja_seg==0,0,mv_mlimitecompra/mcuentas_saldo_mas_invers_mas_caja_seg) ]

    #Productos de dificil desvinculacion
    dataset[ ,   cant_prod_dif_desvinc  := rowSums( cbind(cdebitos_automaticos,cseguro,cinversion12,cprestamos ) , na.rm=TRUE ) ]

    #Mes
    dataset[ , mes   := str_sub(foto_mes, start=-2)  ]
    
  #valvula de seguridad para evitar valores infinitos
  #paso los infinitos a NULOS
  infinitos      <- lapply(names(dataset),function(.name) dataset[ , sum(is.infinite(get(.name)))])
  infinitos_qty  <- sum( unlist( infinitos) )
  if( infinitos_qty > 0 )
  {
    cat( "ATENCION, hay", infinitos_qty, "valores infinitos en tu dataset. Seran pasados a NA\n" )
    dataset[mapply(is.infinite, dataset)] <<- NA
  }


  #valvula de seguridad para evitar valores NaN  que es 0/0
  #paso los NaN a 0 , decision polemica si las hay
  #se invita a asignar un valor razonable segun la semantica del campo creado
  nans      <- lapply(names(dataset),function(.name) dataset[ , sum(is.nan(get(.name)))])
  nans_qty  <- sum( unlist( nans) )
  if( nans_qty > 0 )
  {
    cat( "ATENCION, hay", nans_qty, "valores NaN 0/0 en tu dataset. Seran pasados arbitrariamente a 0\n" )
    cat( "Si no te gusta la decision, modifica a gusto el programa!\n\n")
    dataset[mapply(is.nan, dataset)] <<- 0
  }

}

#primero agrego las variables manuales
AgregarVariables( dataset )

fwrite( dataset,
        file="./datasets/alt_c2_reparacNA_feinicialymio.csv.gz",
        sep= "," )

nrow(dataset)

ncol(dataset)
