==> src/Competencia2/UltimoIntento/5_estrategia_entrenam.ipynb <==
#limpio la memoria
rm( list=ls() )  #remove all objects
gc()             #garbage collection

require("data.table")

#setwd("C:/_MCD/Labo1/code")  #Establezco el Working Directory
setwd("~/buckets/b1/")

kdataset <- "alt_c2_reparacNA_feinicialymioylag1_deflacion.csv.gz"
kexperimento <- "comp2-exp4M01"

k4meses <- TRUE
k3mesesSinDic <- FALSE

if(k4meses)
{
    ktraining <- c( 202010,202011,202012,202101)
    kvalidation <- c( 202102 )
    ktesting <- c( 202103 )
    kfinal_train <- c( 202012,202101,202102,202103 )

    kfuture <- c( 202105 )
}

if(k3mesesSinDic)
{
    ktraining <- c( 202010,202011,202101)
    kvalidation <- c( 202102 )
    ktesting <- c( 202103 )
    kfinal_train <- c( 202101,202102,202103 )

    kfuture <- c( 202105 )
} 

#cargo el dataset
dataset  <- fread(paste0("./datasets/",kdataset) ) 

dir.create( paste0("./exp/",kexperimento), showWarnings = FALSE )


setorder( dataset, foto_mes, numero_de_cliente )

#grabo los datos del futuro
fwrite( dataset[ foto_mes %in% kfuture, ],
        file= paste0("./exp/",kexperimento,"/dataset_future.csv.gz"),
        logical01= TRUE,
        sep= "," )

#grabo los datos donde voy a entrenar los Final Models
fwrite( dataset[ foto_mes %in% kfinal_train, ],
        file= paste0("./exp/",kexperimento,"/dataset_train_final.csv.gz"),
        logical01= TRUE,
        sep= "," )

#grabo los datos donde voy a hacer el training y la optimizacion de hiperparametros
dataset[  , fold_train := 0L ]
dataset[ foto_mes %in% ktraining, fold_train := 1L ]

dataset[  , fold_validate := 0L ]
dataset[ foto_mes %in% kvalidation, fold_validate := 1L ]

dataset[  , fold_test := 0L ]
dataset[ foto_mes %in% ktesting, fold_test := 1L ]

fwrite( dataset[ fold_train + fold_validate + fold_test >= 1 , ],
        file= paste0("./exp/",kexperimento,"/dataset_training.csv.gz"),
        logical01= TRUE,
        sep= "," )
